+++ src/components/day.js
@@ -1,0 +1,16 @@
+import {createTripsEventTemplate} from "./event.js";
+
+const createTripsDayTemplate = () => {
+  return (
+    `<li class="trip-days__item  day">
+      <div class="day__info">
+        <span class="day__counter">1</span>
+        <time class="day__date" datetime="2019-03-18">MAR 18</time>
+      </div>
+    
+      <ul class="trip-events__list"></ul>
+    </li>`
+  );
+};
+
+export {createTripsDayTemplate, createTripsEventTemplate};
