+++ src/components/header.js
@@ -1,0 +1,5 @@
+import {createTripInfoTemplate} from "./trip-info.js";
+import {createMenuTemplate} from "./menu.js";
+import {createFilterTemplate} from "./filter.js";
+
+export {createTripInfoTemplate, createMenuTemplate, createFilterTemplate};
